Added request levels
Necessary to support limited recursion of nested nodes for requests. Instead of inifinite recursion, this provides up to ten levels of depth. The root of the problem is with the Dry Schema/Validation gems which don't support recursion by default.

None of this is ideal but provides a temporary workaround in the interim.